+++ R/anchor.R
@@ -1,6 +1,6 @@
 #' Specify cell limits via an anchor cell
 #'
-#' Specify the targetted cell rectangle via an "upper left" anchor cell and the
+#' Specify the targetted cell rectangle via an upper left anchor cell and the
 #' rectangle's row and column extent. The extent can be specified directly via
 #' \code{dims} or indirectly via the \code{input} object. Specification via
 #' \code{input} anticipates a write operation into the spreadsheet. If
@@ -23,7 +23,7 @@
 #' @param byrow logical, indicating whether a one-dimensional input should run
 #'   down or to the right
 #'
-#' @return a \code{cell_limits} object
+#' @return a \code{\link{cell_limits}} object
 #'
 #' @examples
 #' anchored()
@@ -36,6 +36,7 @@
 #'
 #' anchored(anchor = "R4C2", dim = c(8, 2))
 #' as.range(anchored(anchor = "R4C2", dim = c(8, 2)))
+#' as.range(anchored(anchor = "R4C2", dim = c(8, 2)), RC = TRUE)
 #' dim(anchored(anchor = "R4C2", dim = c(8, 2)))
 #'
 #' (input <- head(iris))
@@ -98,8 +99,8 @@ anchored <- function(anchor = "A1",
     input_extent[1] <- input_extent[1] + 1
   }
 
-  cell_limits(rows = anchorCL$rows + c(0L, input_extent[1] - 1),
-              cols = anchorCL$cols + c(0L, input_extent[2] - 1))
+  cell_limits(ul = anchorCL$ul,
+              lr = anchorCL$lr + input_extent - 1)
 
 }