+++ tests/testthat/test-cell-specification.R
@@ -3,7 +3,7 @@ context("cell specification")
 test_that("Dollar signs are removed", {
 
   expect_equal(rm_dollar_signs(c("A$1:$B$32", "$D11")), c("A1:B32", "D11"))
-  expect_equal(as.cell_limits("A$1:$B$32"), cell_limits(c(1, 32), c(1, 2)))
+  expect_equal(as.cell_limits("A$1:$B$32"), cell_limits(c(1, 1), c(32, 2)))
 
 })
 
@@ -63,7 +63,7 @@ test_that("Cell range is converted to a cell_limit object and vice versa", {
 
   rgA1 <- "A1:C4"
   rgRC <- "R1C1:R4C3"
-  rgCL <- cell_limits(rows = c(1, 4), cols = c(1, 3))
+  rgCL <- cell_limits(ul = c(1, 1), lr = c(4, 3))
   expect_equal(as.cell_limits(rgA1), rgCL)
   expect_equal(as.cell_limits(rgRC), rgCL)
   expect_equal(as.range(rgCL), rgA1)
@@ -73,7 +73,7 @@ test_that("Cell range is converted to a cell_limit object and vice versa", {
   rgA1A1 <- "E7:E7"
   rgRC <- "R7C5"
   rgRCRC <- "R7C5:R7C5"
-  rgCL <- cell_limits(rows = c(7, 7), cols = c(5, 5))
+  rgCL <- cell_limits(ul = c(7, 5), lr = c(7, 5))
   expect_equal(as.cell_limits(rgA1), rgCL)
   expect_equal(as.cell_limits(rgRC), rgCL)
   expect_equal(as.cell_limits(rgA1A1), rgCL)
@@ -81,7 +81,7 @@ test_that("Cell range is converted to a cell_limit object and vice versa", {
   expect_equal(as.range(rgCL), rgA1A1)
   expect_equal(as.range(rgCL, RC = TRUE), rgRCRC)
 
-  rgCL <- cell_limits(rows = c(NA, 4), cols = c(1, NA))
+  rgCL <- cell_limits(ul = c(NA, 1), lr = c(4, NA))
   expect_true(is.na(as.range(rgCL)))
 
 })
@@ -96,9 +96,9 @@ test_that("Bad cell ranges throw errors", {
   expect_error(as.cell_limits("14:17"))
   expect_error(as.cell_limits(14:17))
   expect_error(as.cell_limits(B2:D9))
-  expect_error(cell_limits(rows = c(-1, 3), cols = c(1, 4)))
-  expect_error(cell_limits(rows = c(0, 3), cols = c(1, 4)))
-  expect_error(cell_limits(rows = c(1, 3), cols = c(4, 1)))
+  expect_error(cell_limits(ul = c(-1, 1), lr = c(3, 4)))
+  expect_error(cell_limits(ul = c(0, 1), lr = c(3, 4)))
+  expect_error(cell_limits(ul = c(1, 4), lr = c(3, 1)))
 
 })
 
@@ -106,12 +106,12 @@ test_that("Degenerate, all-NA input is tolerated", {
 
   cl <- cell_limits()
   expect_is(cl, "cell_limits")
-  expect_is(cl$rows, "integer")
+  expect_is(cl$ul, "integer")
 
   cl2 <- cell_limits(c(NA, NA))
   expect_identical(cl, cl2)
 
-  cl3 <- cell_limits(cols = c(NA, NA))
+  cl3 <- cell_limits(lr = c(NA, NA))
   expect_identical(cl, cl3)
 
 })
@@ -131,11 +131,11 @@ test_that("cell_limits objects inherit from list", {
 test_that("Row-only specifications work", {
 
   expect_identical(cell_rows(c(NA, NA)), cell_limits())
-  expect_identical(cell_rows(c(NA, 3)), cell_limits(rows = c(NA, 3)))
-  expect_identical(cell_rows(c(7, NA)), cell_limits(rows = c(7, NA)))
-  expect_identical(cell_rows(c(3, NA, 10)), cell_limits(rows = c(3, 10)))
-  expect_identical(cell_rows(c(10, NA, 3)), cell_limits(rows = c(3, 10)))
-  expect_identical(cell_rows(4:16), cell_limits(rows = c(4L, 16L)))
+  expect_identical(cell_rows(c(NA, 3)), cell_limits(lr = c(3, NA)))
+  expect_identical(cell_rows(c(7, NA)), cell_limits(c(7, NA)))
+  expect_identical(cell_rows(c(3, NA, 10)), cell_limits(c(3, NA), c(10, NA)))
+  expect_identical(cell_rows(c(10, NA, 3)), cell_limits(c(3, NA), c(10, NA)))
+  expect_identical(cell_rows(4:16), cell_limits(c(4, NA), c(16, NA)))
   expect_error(cell_rows(c(7, 2)))
 
 })
@@ -143,16 +143,16 @@ test_that("Row-only specifications work", {
 test_that("Column-only specifications work", {
 
   expect_identical(cell_cols(c(NA, NA)), cell_limits())
-  expect_identical(cell_cols(c(NA, 3)), cell_limits(cols = c(NA, 3)))
-  expect_identical(cell_cols(c(7, NA)), cell_limits(cols = c(7, NA)))
-  expect_identical(cell_cols(c(3, NA, 10)), cell_limits(cols = c(3, 10)))
-  expect_identical(cell_cols(c(10, NA, 3)), cell_limits(cols = c(3, 10)))
-  expect_identical(cell_cols(4:16), cell_limits(cols = c(4L, 16L)))
+  expect_identical(cell_cols(c(NA, 3)), cell_limits(lr = c(NA, 3)))
+  expect_identical(cell_cols(c(7, NA)), cell_limits(c(NA, 7)))
+  expect_identical(cell_cols(c(3, NA, 10)), cell_limits(c(NA, 3), c(NA, 10)))
+  expect_identical(cell_cols(c(10, NA, 3)), cell_limits(c(NA, 3), c(NA, 10)))
+  expect_identical(cell_cols(4:16), cell_limits(c(NA, 4), c(NA, 16)))
   expect_error(cell_cols(c(7, 2)))
 
-  expect_identical(cell_cols("B:D"), cell_limits(cols = c(2L, 4L)))
-  expect_identical(cell_cols(c("C", "ZZ")), cell_limits(cols = c(3L, 702L)))
-  expect_identical(cell_cols(c("C", NA)), cell_limits(cols = c(3L, NA)))
+  expect_identical(cell_cols("B:D"), cell_limits(c(NA, 2), c(NA, 4)))
+  expect_identical(cell_cols(c("C", "ZZ")), cell_limits(c(NA, 3), c(NA, 702)))
+  expect_identical(cell_cols(c("C", NA)), cell_limits(c(NA, 3)))
   expect_error(cell_cols("Z:M"))
   expect_error(cell_cols(c("Z", "M")))
 
@@ -161,7 +161,7 @@ test_that("Column-only specifications work", {
 test_that("Print method works", {
 
   expect_output(cell_limits(c(NA, 7), c(3, NA)),
-                "<cell_limits (-, 3) x (7, -)>", fixed = TRUE)
+                "<cell_limits (1, 7) x (3, -)>", fixed = TRUE)
 
 })
 
@@ -169,11 +169,10 @@ test_that("dim method works", {
 
   expect_equivalent(dim(as.cell_limits("A1")), c(1, 1))
   expect_equivalent(dim(as.cell_limits("A1:F10")), c(10, 6))
-  expect_equivalent(dim(cell_limits(c(1, 1), c(2, 5))), c(1, 4))
-  expect_equivalent(dim(cell_limits(c(NA, 1), c(2, 5))), c(NA_integer_, 4))
-  expect_equivalent(dim(cell_limits(c(NA, NA), c(2, 5))), c(NA_integer_, 4))
-  expect_equivalent(dim(cell_limits(c(NA, 1), c(NA, 5))),
-                    c(NA_integer_, NA_integer_))
+  expect_equivalent(dim(cell_limits(c(1, 2), c(1, 5))), c(1, 4))
+  expect_equivalent(dim(cell_limits(c(NA, 2), c(1, 5))), c(1, 4))
+  expect_equivalent(dim(cell_limits(c(NA, 2), c(NA, 5))), c(NA_integer_, 4))
+  expect_equivalent(dim(cell_limits(c(1, 1))), c(NA_integer_, NA_integer_))
 
 })
 
@@ -182,35 +181,35 @@ test_that("Cell limits can be specified via anchor", {
   ## no input
   expect_identical(anchored(), as.cell_limits("A1"))
   expect_identical(anchored(anchor = "R4C2", dim = c(8, 2)),
-                   cell_limits(c(4, 11), c(2, 3)))
+                   cell_limits(c(4, 2), c(11, 3)))
   expect_identical(anchored(anchor = "A1", dim = c(3, 3), col_names = FALSE),
-                   cell_limits(c(1, 3), c(1, 3)))
+                   cell_limits(c(1, 1), c(3, 3)))
   expect_identical(anchored(anchor = "A1", dim = c(3, 3), col_names = TRUE),
-                   cell_limits(c(1, 4), c(1, 3)))
+                   cell_limits(c(1, 1), c(4, 3)))
 
   ## 2-dimensional input
   input <- head(iris)
   expect_identical(anchored(anchor = "R3C7", input = input),
-                   cell_limits(c(3, 9), c(7, 11)))
+                   cell_limits(c(3, 7), c(9, 11)))
   expect_identical(anchored(anchor = "R3C7", input = input, col_names = TRUE),
-                   cell_limits(c(3, 9), c(7, 11)))
+                   cell_limits(c(3, 7), c(9, 11)))
   expect_identical(anchored(anchor = "R3C7", input = input, col_names = FALSE),
-                   cell_limits(c(3, 8), c(7, 11)))
+                   cell_limits(c(3, 7), c(8, 11)))
   ## dim should have no effect here
   expect_identical(anchored(anchor = "R3C7", input = input, dim = c(2,2)),
-                   cell_limits(c(3, 9), c(7, 11)))
+                   cell_limits(c(3, 7), c(9, 11)))
 
   ## 1-dimensional input
   input <- LETTERS[1:8]
   expect_identical(anchored(anchor = "B5", input = input),
-                   cell_limits(c(5, 12), c(2, 2)))
+                   cell_limits(c(5, 2), c(12, 2)))
   expect_identical(anchored(anchor = "B5", input = input, byrow = TRUE),
-                   cell_limits(c(5, 5), c(2, 9)))
+                   cell_limits(c(5, 2), c(5, 9)))
   ## dim and col_names should have no effect here
   expect_identical(anchored(anchor = "B5", input = input, dim = c(5,5)),
-                   cell_limits(c(5, 12), c(2, 2)))
+                   cell_limits(c(5, 2), c(12, 2)))
   expect_identical(anchored(anchor = "B5", input = input, col_names = TRUE),
-                   cell_limits(c(5, 12), c(2, 2)))
+                   cell_limits(c(5, 2), c(12, 2)))
 
   expect_error(anchored(1))
   expect_error(anchored("A"))